ContainerTheme

Description

This module is part of StrategyLib, a library of functional strategy combinators, including combinators for generic traversal. This module provides combinators which allow one to use strategies to construct generic containers.

Pointwise function update

modify :: GHC.Classes.Eq x => (x -> y) -> x -> y -> x -> y

Pointwise modification of monomorphic functions

modifyTP :: (Control.Monad.MonadPlus m, GHC.Classes.Eq t, TermRep.Term t) => StrategyPrimitives.TP m -> t -> m t -> StrategyPrimitives.TP m

Pointwise modification of type-preserving strategies

modifyTU :: (Control.Monad.MonadPlus m, GHC.Classes.Eq t, TermRep.Term t) => StrategyPrimitives.TU a m -> t -> m a -> StrategyPrimitives.TU a m

Pointwise modification of type-unifying strategies

Generic Set (not observable)

type GSet = StrategyPrimitives.TU GHC.Tuple.() Data.Maybe.Maybe

Type of generic sets

emptyGSet :: ContainerTheme.GSet

Empty generic set.

fullGSet :: ContainerTheme.GSet

Completely filled generic set

addGSet :: (GHC.Classes.Eq t, TermRep.Term t) => t -> ContainerTheme.GSet -> ContainerTheme.GSet

Add an element to a generic set

removeGSet :: (GHC.Classes.Eq t, TermRep.Term t) => t -> ContainerTheme.GSet -> ContainerTheme.GSet

Remove an element from a generic set

containsGSet :: (GHC.Classes.Eq t, TermRep.Term t) => t -> ContainerTheme.GSet -> GHC.Types.Bool

Test whether a given element is contained in a generic set

Generic Map (not observable)

type GMap value = StrategyPrimitives.TU value Data.Maybe.Maybe

Type of generic maps

emptyGMap :: ContainerTheme.GMap v

Empty generic map

removeGMap :: (GHC.Classes.Eq t, TermRep.Term t) => t -> ContainerTheme.GMap v -> ContainerTheme.GMap v

Remove an element from a generic map (my key)

containsGMap :: (GHC.Classes.Eq t, TermRep.Term t) => t -> ContainerTheme.GMap v -> GHC.Types.Bool

Test whether an element with given key is contained in a generic map

putGMap :: (GHC.Classes.Eq t, TermRep.Term t) => t -> v -> ContainerTheme.GMap v -> ContainerTheme.GMap v

Add an entry with given key and value to a generic map

getGMap :: (GHC.Classes.Eq t, TermRep.Term t) => t -> ContainerTheme.GMap v -> Data.Maybe.Maybe v

Obtain the value for a given key from a generic map

type GList = (GHC.Integer.Type.Integer -> StrategyPrimitives.TP Data.Maybe.Maybe, GHC.Integer.Type.Integer)

sizeGList :: (t, t1) -> t1

indxGList :: (t, t1) -> t

emptyGList :: ContainerTheme.GList

addGList :: TermRep.Term t => t -> ContainerTheme.GList -> ContainerTheme.GList

putGList :: TermRep.Term t => GHC.Integer.Type.Integer -> t -> ContainerTheme.GList -> ContainerTheme.GList

getGList :: TermRep.Term t => GHC.Integer.Type.Integer -> ContainerTheme.GList -> Data.Maybe.Maybe t

mapGListTP :: StrategyPrimitives.TP Data.Maybe.Maybe -> ContainerTheme.GList -> ContainerTheme.GList

mapGListTU :: TermRep.Term t => (t -> GHC.Tuple.()) -> StrategyPrimitives.TU a Data.Maybe.Maybe -> ContainerTheme.GList -> [Data.Maybe.Maybe a]

elemsGList :: TermRep.Term t => (t -> GHC.Tuple.()) -> ContainerTheme.GList -> [t]

nth :: [a] -> GHC.Integer.Type.Integer -> a

type Coder = (GHC.Types.Int, StrategyPrimitives.TU GHC.Types.Int Data.Maybe.Maybe)

noCode :: ContainerTheme.Coder

getCode :: TermRep.Term x => ContainerTheme.Coder -> x -> Data.Maybe.Maybe GHC.Types.Int

setCode :: (TermRep.Term x, GHC.Classes.Eq x) => ContainerTheme.Coder -> x -> GHC.Types.Int -> ContainerTheme.Coder

nextCode :: ContainerTheme.Coder -> (GHC.Types.Int, ContainerTheme.Coder)

enCode :: (TermRep.Term x, GHC.Classes.Eq x) => ContainerTheme.Coder -> x -> ContainerTheme.Coder