EffectTheme

Description

This module is part of StrategyLib, a library of functional strategy combinators, including combinators for generic traversal. This module provides combinators to localize monadic effects.

Replace one strategy monad by another

mrunTP :: (GHC.Base.Monad m, GHC.Base.Monad m', MonadRun.MonadRun s m) => (forall a. s a a) -> StrategyPrimitives.TP m -> StrategyPrimitives.TP m'

Replace the monad in a type-preserving strategy, given a monad algebra (see MonadRun.MonadRun) for the monad that is replaced. The two monads are unrelated, so none of the effects in the monad that is replaced carry over to the one that replaces it.

mrunTU :: (GHC.Base.Monad m, GHC.Base.Monad m', MonadRun.MonadRun s m) => s a a -> StrategyPrimitives.TU a m -> StrategyPrimitives.TU a m'

Replace the monad in a type-unifying strategy, given a monad algebra (see MonadRun.MonadRun) for the monad that is replaced. The two monads are unrelated, so none of the effects in the monad that is replaced carry over to the one that replaces it.

Add an effect to the strategy monad

liftTP :: (GHC.Base.Monad (t m), GHC.Base.Monad m, Control.Monad.Trans.Class.MonadTrans t) => StrategyPrimitives.TP m -> StrategyPrimitives.TP (t m)

Add an effect to the monad in a type-preserving strategy. The monads are related by a monad transformer, so the effects of the incoming monad are preserved in the result monad. We use the Control.Monad.Trans.Class.lift function of the monad transformer.

liftTU :: (GHC.Base.Monad (t m), GHC.Base.Monad m, Control.Monad.Trans.Class.MonadTrans t) => StrategyPrimitives.TU a m -> StrategyPrimitives.TU a (t m)

Add an effect to the monad in a type-unifying strategy. The monads are related by a monad transformer, so the effects of the incoming monad are preserved in the result monad. We use the Control.Monad.Trans.Class.lift function of the monad transformer.

Remove an effect from the strategy monad

unliftTP :: (GHC.Base.Monad (t m), GHC.Base.Monad m, MonadRun.MonadUnTrans s t) => (forall a. s a a) -> StrategyPrimitives.TP (t m) -> StrategyPrimitives.TP m

remove an effect from the monad of a type-preserving strategy. The monads are related by a monad untransformer (see MonadRun.MonadUnTrans), so the effects of the incoming monad are preserved in the result monad, except for the effect for which a monad algebra is supplied.

unliftTU :: (GHC.Base.Monad (t m), GHC.Base.Monad m, MonadRun.MonadUnTrans s t) => s a a -> StrategyPrimitives.TU a (t m) -> StrategyPrimitives.TU a m

remove an effect from the monad of a type-unifying strategy. The monads are related by a monad untransformer (see MonadRun.MonadUnTrans), so the effects of the incoming monad are preserved in the result monad, except for the effect for which a monad algebra is supplied.

Localize specific effects

Localize the partiality effect

guaranteeSuccessTP

Arguments

:: (GHC.Base.Monad (t m), GHC.Base.Monad m, MonadRun.MonadUnTrans MonadRun.MaybeAlg t)

=> (forall a. a)
default value (Note: universally quantified!)

-> StrategyPrimitives.TP (t m)
type-preserving partial strategy

-> StrategyPrimitives.TP m
type-preserving strategy without partiality

Localize the partiality effect in a type-preserving strategy. A default value must be supplied to be used to recover from failure. Since this default parameter is universally quantified, only GHC.Err.undefined and 'error ...' can be used to instantiate it. See also 'unsafeGuaranteeSuccessTP.

guaranteeSuccessTU

Arguments

:: (GHC.Base.Monad (t m), GHC.Base.Monad m, MonadRun.MonadUnTrans MonadRun.MaybeAlg t)

=> a
default value

-> StrategyPrimitives.TU a (t m)
type-preserving partial strategy

-> StrategyPrimitives.TU a m
type-preserving strategy without partiality

Localize the partiality effect in a type-unifying strategy. A default value must be supplied to be used to recover from failure.

unsafeGuaranteeSuccessTP :: (GHC.Base.Monad (t m), GHC.Base.Monad m, MonadRun.MonadUnTrans MonadRun.MaybeAlg t) => StrategyPrimitives.TP (t m) -> StrategyPrimitives.TP m

Unsafe version of EffectTheme.guaranteeSuccessTP. This version uses uses GHC.Err.undefined to recover from failure. For the type-preserving case, this is the only possible default value.

Localize the state effect

localStateTP :: (GHC.Base.Monad (t m), GHC.Base.Monad m, MonadRun.MonadUnTrans (MonadRun.StateAlg s) t) => s -> StrategyPrimitives.TP (t m) -> StrategyPrimitives.TP m

Localize the state of a type-preserving strategy. The first argument represents the initial state.

localStateTU :: (GHC.Base.Monad (t m), GHC.Base.Monad m, MonadRun.MonadUnTrans (MonadRun.StateAlg s) t) => s -> StrategyPrimitives.TU a (t m) -> StrategyPrimitives.TU a m

Localize the state of a type-unifying strategy. The first argument represents the initial state.