FixpointTheme

Description

This module is part of StrategyLib, a library of functional strategy combinators, including combinators for generic traversal. This module defines combinators that iterate until some kind of fixpoint is reached.

Fixpoint Iteration

repeatTP :: Control.Monad.MonadPlus m => StrategyPrimitives.TP m -> StrategyPrimitives.TP m

Exhaustive repeated application at the root of the input term

Fixpoint Traversal

reduce :: Control.Monad.MonadPlus m => StrategyPrimitives.TP m -> StrategyPrimitives.TP m

Exhaustive repeated application throughout the input term.

outermost :: Control.Monad.MonadPlus m => StrategyPrimitives.TP m -> StrategyPrimitives.TP m

Exhaustive repeated application according to the left-most outermost traversal strategy.

innermost' :: Control.Monad.MonadPlus m => StrategyPrimitives.TP m -> StrategyPrimitives.TP m

Exhaustive repeated application according to the left-most innermost traversal strategy, implemented in a naive way. Use FixpointTheme.innermost instead.

innermost :: Control.Monad.MonadPlus m => StrategyPrimitives.TP m -> StrategyPrimitives.TP m

Exhaustive repeated application according to the left-most innermost traversal strategy, implemented in a more efficient way.