FlowTheme

Description

This module is part of StrategyLib, a library of functional strategy combinators, including combinators for generic traversal. This module defines combinators to wire up control and data flow. Whenever possible, we define the combinators in an overloaded fashion but we provide type-specialised variants for TP and TU for convenience.

Try: recover from failure

tryS :: (OverloadingTheme.StrategyPlus s m, OverloadingTheme.StrategyMonoid s m) => s m -> s m

Attempt a strategy s, but recover if it fails.

tryTP :: Control.Monad.MonadPlus m => StrategyPrimitives.TP m -> StrategyPrimitives.TP m

Attempt a type-preserving strategy s, but if it fails, return the input term unchanged.

tryTU :: (Control.Monad.MonadPlus m, Data.Monoid.Monoid u) => StrategyPrimitives.TU u m -> StrategyPrimitives.TU u m

Attempt a type-unifying strategy s, but if it fails, return the Data.Monoid.mempty element of a Data.Monoid.Monoid.

Test: ignore result, but retain effects

testS :: OverloadingTheme.Strategy s m => s m -> StrategyPrimitives.TP m

Test for a strategy's success in a type-preserving context.

testTP :: GHC.Base.Monad m => StrategyPrimitives.TP m -> StrategyPrimitives.TP m

Test for a type-preserving strategy's success in a type-preserving context.

testTU :: GHC.Base.Monad m => StrategyPrimitives.TU a m -> StrategyPrimitives.TP m

Test for a type-unifying strategy's success in a type-preserving context.

If-then-else: pass value from condition to then-clause

ifS :: OverloadingTheme.StrategyPlus s m => StrategyPrimitives.TU u m -> (u -> s m) -> s m -> s m

If c succeeds, pass its value to the then-clause t, otherwise revert to the else-clause e.

ifTP :: Control.Monad.MonadPlus m => StrategyPrimitives.TU u m -> (u -> StrategyPrimitives.TP m) -> StrategyPrimitives.TP m -> StrategyPrimitives.TP m

If c succeeds, pass its value to the then-clause t, otherwise revert to the else-clause e.

ifTU :: Control.Monad.MonadPlus m => StrategyPrimitives.TU u m -> (u -> StrategyPrimitives.TU u' m) -> StrategyPrimitives.TU u' m -> StrategyPrimitives.TU u' m

If c succeeds, pass its value to the then-clause t, otherwise revert to the else-clause e.

If-then: disciplined form of a guarding

ifthenS :: OverloadingTheme.Strategy s m => StrategyPrimitives.TU GHC.Tuple.() m -> s m -> s m

Guard then-clause t by the void-valued type-unifying condition c.

ifthenTP :: GHC.Base.Monad m => StrategyPrimitives.TU GHC.Tuple.() m -> StrategyPrimitives.TP m -> StrategyPrimitives.TP m

Guard type-preserving then-clause t by the void-valued type-unifying condition c.

ifthenTU :: GHC.Base.Monad m => StrategyPrimitives.TU GHC.Tuple.() m -> StrategyPrimitives.TU u m -> StrategyPrimitives.TU u m

Guard type-unifying then-clause t by the void-valued type-unifying condition c.

Not: negation by failure

notS :: OverloadingTheme.StrategyPlus s m => s m -> StrategyPrimitives.TP m

Invert the success-value of strategy s.

notTP :: Control.Monad.MonadPlus m => StrategyPrimitives.TP m -> StrategyPrimitives.TP m

Invert the success-value of type-preserving strategy s. Its output term (in case of success) will be ignored.

notTU :: Control.Monad.MonadPlus m => StrategyPrimitives.TU u m -> StrategyPrimitives.TP m

Invert the success-value of type-unifying strategy s. Its output value (in case of success) will be ignored.

Exclusive choice

xchoiceS :: OverloadingTheme.StrategyPlus s m => s m -> s m -> s m

Succeed if exactly one argument strategy succeeds.

xchoiceTP :: Control.Monad.MonadPlus m => StrategyPrimitives.TP m -> StrategyPrimitives.TP m -> StrategyPrimitives.TP m

Succeed if exactly one argument strategy succeeds.

xchoiceTU :: Control.Monad.MonadPlus m => StrategyPrimitives.TU u m -> StrategyPrimitives.TU u m -> StrategyPrimitives.TU u m

Succeed if exactly one argument strategy succeeds.

Generic filter, derived from monomorphic predicate

filterTP :: (TermRep.Term t, Control.Monad.MonadPlus m) => (t -> GHC.Types.Bool) -> StrategyPrimitives.TP m

If predicate g holds for the input term, return it as output term, otherwise fail.

filterTU :: (TermRep.Term t, Control.Monad.MonadPlus m) => (t -> GHC.Types.Bool) -> StrategyPrimitives.TU t m

If predicate g holds for the input term, return it as output value, otherwise fail.

Generic ticker, derived from monomorphic predicate

tickTU :: (GHC.Base.Monad m, TermRep.Term t, GHC.Num.Num n) => (t -> GHC.Types.Bool) -> StrategyPrimitives.TU n m

If predicate g holds for the input term, return 1 otherwise return 0.

Type guards

type TypeGuard a = a -> GHC.Tuple.()

Type guard (function type), i.e., guard that does not observe values

typeGuard :: FlowTheme.TypeGuard a

Type guard (function). Typical usage:

 
   full_tdTU (typeTickTU (typeGuard::TypeGuard MyType))

Generic ticker, derived from type guard

typeTickTU :: (TermRep.Term t, GHC.Base.Monad m, GHC.Num.Num n) => FlowTheme.TypeGuard t -> StrategyPrimitives.TU n m

If type guard holds for the input term, return 1 otherwise return 0.

Generic filters, derived from type guard

typeFilterTP :: (TermRep.Term t, Control.Monad.MonadPlus m) => FlowTheme.TypeGuard t -> StrategyPrimitives.TP m

If type guard holds for the input term, return it as output term, otherwise fail.

typeFilterTU :: (TermRep.Term t, Control.Monad.MonadPlus m) => FlowTheme.TypeGuard t -> StrategyPrimitives.TU t m

If type guard holds for the input term, return it as output value, otherwise fail.