KeyholeTheme

Description

This module is part of StrategyLib, a library of functional strategy combinators, including combinators for generic traversal. This module defines a number combinators for keyhole operations, i.e. for operations that have ordinary parametric or adhoc polymorhpic types, but employ strategies inside.

Focus

selectFocus

Arguments

:: (TermRep.Term f, TermRep.Term t)

=> (f -> Data.Maybe.Maybe f)
Identify focus

-> t
Input term

-> Data.Maybe.Maybe f
Focused term

Select the identified focus. Fails if no subterm can be selected.

replaceFocus :: (TermRep.Term t, TermRep.Term t') => (t -> Data.Maybe.Maybe t) -> t' -> Data.Maybe.Maybe t'

Replace the identified focus. Fails if no subterm can be replaced.

deleteFocus

Arguments

:: (TermRep.Term f, TermRep.Term t)

=> (f -> Data.Maybe.Maybe f)
Identify focus

-> t
Input term

-> Data.Maybe.Maybe t
Output term without focused entity

Delete the focus assuming it is an element in a list. Fails if no deletion can be performed.

selectHost

Arguments

:: (TermRep.Term f, TermRep.Term h, TermRep.Term t)

=> (f -> Data.Maybe.Maybe f)
Get focus

-> (h -> Data.Maybe.Maybe h)
Get host

-> t
Input term

-> Data.Maybe.Maybe h
Located host

Find the host of the focused entity, i.e. a superterm of the focussed subterm.

markHost

Arguments

:: (TermRep.Term f, TermRep.Term h, TermRep.Term t)

=> (f -> GHC.Types.Bool)
Test focus

-> (h -> h)
Wrap host

-> t
Input term

-> Data.Maybe.Maybe t
Output term

Listification

listify :: (TermRep.Term x, TermRep.Term y) => x -> [y]

Put all nodes of a certain type into a list.

strings :: TermRep.Term x => x -> [GHC.Base.String]

Put all nodes of type GHC.Base.String into a list. This is a type-specialization of KeyholeTheme.listify.

Keyhole versions of basic strategy combinators.

inj :: (Control.Monad.MonadPlus m, TermRep.Term x, TermRep.Term c) => (c -> m c) -> x -> m x

Apply the argument function to the unique subterm of the input term. Fail if the input term has more subterms or if the subterm is not of the appropriate type. This is a keyhole version of the traversal combinator StrategyPrimitives.injTP