MetricsTheme

Description

This module is part of StrategyLib, a library of functional strategy combinators, including combinators for generic traversal. This module defines combinators to define metrics extractors.

An abstract datatype for metrics

type Metrics = MetricsTheme.MetricName -> GHC.Integer.Type.Integer

The type of metrics

type MetricName = GHC.Base.String

The type of metric names

initMetrics :: GHC.Integer.Type.Integer -> MetricsTheme.Metrics

Create MetricsTheme.Metrics with given initial value for all metrics.

initMetrics0 :: MetricsTheme.Metrics

Create MetricsTheme.Metrics with 0 as initial value for all metrics.

incMetrics :: MetricsTheme.MetricName -> GHC.Integer.Type.Integer -> MetricsTheme.Metrics -> MetricsTheme.Metrics

Create MetricsTheme.Metrics with initTypeMetrics :: MetricName -> a -> Metrics initTypeMetrics key _ = incMetrics1 key initMetrics0

Increment metric with the given name with the given value.

incMetrics1 :: MetricsTheme.MetricName -> MetricsTheme.Metrics -> MetricsTheme.Metrics

Increment metric with the given name by 1.

putMetricLn :: MetricsTheme.MetricName -> MetricsTheme.Metrics -> GHC.Types.IO GHC.Tuple.()

Print value of metric with the given name.

Metrics as monoids

Strategy combinators for metrics

Generic metric algorithms

depthWith

Arguments

:: Control.Monad.MonadPlus m

=> StrategyPrimitives.TU GHC.Tuple.() m
Recognize relevant contructs

-> StrategyPrimitives.TU GHC.Types.Int m
Count nesting depth of relevant constructs.

Generic algorithm for computing nesting depth