MonadRun

Description

This module is part of StrategyLib, a library of functional strategy combinators, including combinators for generic traversal. This module provides non-strategic functionality for running monads and unlifting monad transformers. In a sense, this is dual to the GHC.Base.return and Control.Monad.Trans.Class.lift functionality of the GHC.Base.Monad and Control.Monad.Trans.Class.MonadTrans classes.

Monad algebras

data MaybeAlg a b

The algebra for the partiality effect of Data.Maybe.Maybe and MonadMaybe.MaybeT.

Constructors

MaybeAlg

Fields

nothing :: b

just :: a -> b

data ErrorAlg e a b

The algebra for the error effect of Data.Either.Either and Control.Monad.Trans.Error.ErrorT.

Constructors

ErrorAlg

Fields

left :: e -> b

right :: a -> b

data ListAlg a b

The algebra for the non-determinacy effect of '[]' and Control.Monad.Trans.List.ListT.

Constructors

ListAlg

Fields

nil :: b

cons :: a -> b -> b

data StateAlg s a b

The algebra for the state effect of StrategyLib.State and StrategyLib.StateT.

Constructors

StateAlg

Fields

first :: s

initial state

next :: (a, s) -> b

state transformer

Running monads

class MonadRun s m | m -> s where

The class of monads for which a MonadRun.run function is defined that executes the computation of the monad.

Methods

run :: s a b -> m a -> b

The overloaded function run takes as first argument an algebra which captures the ingredients necessary to run the particular monad at hand. This algebra is parameterized with the domain and co-domain of run.

mrun :: (MonadRun.MonadRun s m, GHC.Base.Monad m') => s a b -> m a -> m' b

Exchange one monad by another. This function runs one monad, and puts its value in another. This is basically a monadic version of the MonadRun.run function itself. Note that the two monads are unrelated, so none of the effects of the incoming monad are transferred to the result monad.

Unlifting monad transformers

class MonadUnTrans s t | t -> s where

Just as a base monad can be run to remove the monad, so can a transformed monad be unlifted to remove the transformer and obtain the original monad.

Methods

unlift :: GHC.Base.Monad m => s a b -> t m a -> m b

The overloaded function MonadRun.unlift for monad transformers takes as first argument an algebra just like the run function for base monads. For each monad transformer, the same algebra is used as for the base monad of which the transformer is the parameterized variant.

Monadic choice combinators that confine the partiality effect

Monadic choice

mplus' :: (GHC.Base.Monad m, MonadRun.MonadUnTrans MonadRun.MaybeAlg t) => t m b -> m b -> m b

Monadic choice combinator that confines the partiality effect to the first argument. This is a variation on Control.Monad.mplus which allows the partiality effect to spread to both arguments and to the result.

mswitch

Arguments

:: (GHC.Base.Monad m, MonadRun.MonadUnTrans MonadRun.MaybeAlg t)

=> [t m b]
choice branches

-> m b
otherwise

-> m b
result

Monadic choice combinator. Generalization of MonadRun.mplus' that takes a list of choice arguments rather than a single one.

mayswitch :: GHC.Base.Monad m => [MonadMaybe.MaybeT m b] -> m b -> m b

Specialization of MonadRun.mswitch for MaybeT.

Monadic function choice

mchoice' :: (GHC.Base.Monad m, MonadRun.MonadUnTrans MonadRun.MaybeAlg t) => (a -> t m b) -> (a -> m b) -> a -> m b

Monadic function choice combinator that confines the partiality effect to the first argument. This is a variation on mchoice which allows the partiality effect to spread to both arguments and to the result.

mchoices :: (GHC.Base.Monad m, MonadRun.MonadUnTrans MonadRun.MaybeAlg t, Control.Monad.MonadPlus (t m)) => [a -> t m b] -> (a -> m b) -> a -> m b

Monadic function choice combinator. Generalization of MonadRun.mchoice' that takes a list of choice arguments rather than a single one.

Implementation variants

mswitch0 :: (GHC.Base.Monad m, MonadRun.MonadUnTrans MonadRun.MaybeAlg t) => [t m b] -> m b -> m b

Implementation variant of MonadRun.mswitch in terms of foldr.

mswitch1 :: (GHC.Base.Monad m, MonadRun.MonadUnTrans MonadRun.MaybeAlg t) => [t m b] -> m b -> m b

Implementation variant of MonadRun.mswitch with MonadRun.mplus' expanded:

mswitch' :: (GHC.Base.Monad m, MonadRun.MonadUnTrans MonadRun.MaybeAlg t, Control.Monad.MonadPlus (t m)) => [t m b] -> m b -> m b

Implementation variant of MonadRun.mswitch where the unlift is postponed to the very end.