MonadicFunctions

Description

This module is part of StrategyLib, a library of functional strategy combinators, including combinators for generic traversal. This module defines auxilliary monadic functions, some of which serve as parametric polymorphic prototypes for actual strategy combinators.

The identity monad

newtype Id a

Identity type constructor.

Constructors

Id a

unId :: MonadicFunctions.Id a -> a

Unwrap identity type constructor

Recover from partiality

succeed :: Data.Maybe.Maybe x -> x

Force success. If the argument value corresponds to failure, a run-time error will occur.

Prototypes for strategy combinators seq, let, choice

mseq :: GHC.Base.Monad m => (a -> m b) -> (b -> m c) -> a -> m c

Sequential composition of monadic functions

mlet :: GHC.Base.Monad m => (a -> m b) -> (b -> a -> m c) -> a -> m c

Sequential composition with value passing; a kind of monadic let.

mchoice :: Control.Monad.MonadPlus m => (a -> m b) -> (a -> m b) -> a -> m b

Choice combinator for monadic functions

Guards and conditionals

argtype :: Control.Monad.MonadPlus m => (x -> y) -> x -> m GHC.Tuple.()

Type guard described by the argument type of a function.

valtype :: Control.Monad.MonadPlus m => x -> x -> m GHC.Tuple.()

Type guard described by a type of a value.

ifM :: Control.Monad.MonadPlus m => m a -> (a -> m c) -> m c -> m c

A kind of monadic conditional.