NameTheme

Description

This module is part of StrategyLib, a library of functional strategy combinators, including combinators for generic traversal. This module provides algorithms to collect names and their types.

Free name analysis

freeNames

Arguments

:: (GHC.Classes.Eq name, TermRep.Term t)

=> StrategyPrimitives.TU [(name, tpe)] StrategyLib.Identity
Identify declarations

-> StrategyPrimitives.TU [name] StrategyLib.Identity
Identify references

-> t
Input term

-> [name]
Free names

Generic free name analysis algorithm (without types)

freeTypedNames

Arguments

:: (GHC.Classes.Eq name, TermRep.Term t)

=> StrategyPrimitives.TU [(name, tpe)] StrategyLib.Identity
Identify declarations

-> StrategyPrimitives.TU [name] StrategyLib.Identity
Identify references

-> [(name, tpe)]
Derived declarations

-> t
Input term

-> [(name, tpe)]
Free names with types

Generic free name analysis algorithm with types

Bound name analysis

boundTypedNames :: (TermRep.Term f, TermRep.Term t, GHC.Classes.Eq name) => StrategyPrimitives.TU [(name, tpe)] StrategyLib.Identity -> (f -> Data.Maybe.Maybe f) -> t -> Data.Maybe.Maybe ([(name, tpe)], f)

Accumulate declarations for focus