PathTheme

Description

This module is part of StrategyLib, a library of functional strategy combinators, including combinators for generic traversal. In this module, we define path combinator to constrain selection and transformation of nodes or subtrees by path conditions.

Below

Strictly below

belowS :: (Control.Monad.MonadPlus m, OverloadingTheme.Strategy s m, OverloadingTheme.StrategyPlus s m) => s m -> StrategyPrimitives.TU GHC.Tuple.() m -> s m

Select or transform a node below a node where a condition holds. We find the top-most node which admits selection or transformation below the top-most node which meets the condition. Thus, the distance between guard and application node is minimized.

Below or at the same height

beloweqS :: (Control.Monad.MonadPlus m, OverloadingTheme.Strategy s m, OverloadingTheme.StrategyPlus s m) => s m -> StrategyPrimitives.TU GHC.Tuple.() m -> s m

Select or transform a node below or at a node where a condition holds.

Type-specialised versions

belowTP :: Control.Monad.MonadPlus m => StrategyPrimitives.TP m -> StrategyPrimitives.TU GHC.Tuple.() m -> StrategyPrimitives.TP m

Apply a transformation strictly below a node where a condition holds.

beloweqTP :: Control.Monad.MonadPlus m => StrategyPrimitives.TP m -> StrategyPrimitives.TU GHC.Tuple.() m -> StrategyPrimitives.TP m

Apply a transformation below or at a node where a condition holds.

Above

Strictly above

aboveS :: (Control.Monad.MonadPlus m, OverloadingTheme.Strategy s m, OverloadingTheme.StrategyPlus s m) => s m -> StrategyPrimitives.TU GHC.Tuple.() m -> s m

Select or transform a node above a node where a condition holds. The distance between guard and application node is minimized.

Above or at the same height

aboveeqS :: (Control.Monad.MonadPlus m, OverloadingTheme.Strategy s m, OverloadingTheme.StrategyPlus s m) => s m -> StrategyPrimitives.TU GHC.Tuple.() m -> s m

Select or transform a node above or at a node where a condition holds.

Type-specialised versions

aboveTP :: Control.Monad.MonadPlus m => StrategyPrimitives.TP m -> StrategyPrimitives.TU GHC.Tuple.() m -> StrategyPrimitives.TP m

Apply a transformation strictly above a node where a condition holds.

aboveeqTP :: Control.Monad.MonadPlus m => StrategyPrimitives.TP m -> StrategyPrimitives.TU GHC.Tuple.() m -> StrategyPrimitives.TP m

Apply a transformation above or at a node where a condition holds.