RefactoringTheme

Description

This module is part of StrategyLib, a library of functional strategy combinators, including combinators for generic traversal. This module defines generic refactoring functionality. See the paper Towards Generic Refactoring by Ralf Laemmel. See also generic-refactoring in the examples directory.

The abstraction interface

class (TermRep.Term abstr, GHC.Classes.Eq name, TermRep.Term [abstr], TermRep.Term apply) => Abstraction abstr name tpe apply | abstr -> name, abstr -> tpe, abstr -> apply, apply -> name, apply -> abstr where

Class of abstractions

Methods

getAbstrName :: abstr -> Data.Maybe.Maybe name

getAbstrParas :: abstr -> Data.Maybe.Maybe [(name, tpe)]

getAbstrBody :: abstr -> Data.Maybe.Maybe apply

getApplyName :: apply -> Data.Maybe.Maybe name

getApplyParas :: apply -> Data.Maybe.Maybe [(name, tpe)]

constrAbstr :: name -> [(name, tpe)] -> apply -> Data.Maybe.Maybe abstr

constrApply :: name -> [(name, tpe)] -> Data.Maybe.Maybe apply

Removal

eliminate

Arguments

:: (TermRep.Term prog, RefactoringTheme.Abstraction abstr name tpe apply)

=> StrategyPrimitives.TU [(name, tpe)] StrategyLib.Identity
Identify declarations

-> StrategyPrimitives.TU [name] StrategyLib.Identity
Identify references

-> (abstr -> Data.Maybe.Maybe abstr)
Unwrap abstraction

-> prog
Input program

-> Data.Maybe.Maybe prog
Output program

Remove an unused abstraction

Insertion

introduce

Arguments

:: (TermRep.Term prog, RefactoringTheme.Abstraction abstr name tpe apply)

=> StrategyPrimitives.TU [(name, tpe)] StrategyLib.Identity
Identify declarations

-> StrategyPrimitives.TU [name] StrategyLib.Identity
Identify references

-> ([abstr] -> Data.Maybe.Maybe [abstr])
Unwrap scope with abstractions

-> abstr
Abstraction to be inserted

-> prog
Input program

-> Data.Maybe.Maybe prog
Output program

Insert a new abstraction

Generic extraction (say fold)

extract

Arguments

:: (TermRep.Term prog, RefactoringTheme.Abstraction abstr name tpe apply)

=> StrategyPrimitives.TU [(name, tpe)] StrategyLib.Identity
Identify declarations

-> StrategyPrimitives.TU [name] StrategyLib.Identity
Identify references

-> (apply -> Data.Maybe.Maybe apply)
Unwrap focus

-> ([abstr] -> [abstr])
Wrap host

-> ([abstr] -> Data.Maybe.Maybe [abstr])
Unwrap host

-> ([(name, tpe)] -> apply -> GHC.Types.Bool)
Check focus

-> name
Name for abstraction

-> prog
Input program

-> Data.Maybe.Maybe prog
Output program

Extract an abstraction