StrategyLib

Description

This module is part of StrategyLib, a library of functional strategy combinators, including combinators for generic traversal. This is the top-level module of the library. One only needs to import this module to use the entire library. Some base modules are exported as well because they are commonly used.

Documentation

newtype Identity a

Constructors

Identity

Fields

runIdentity :: a

type State s = StrategyLib.StateT s StrategyLib.Identity

newtype StateT s m a

Constructors

StateT

Fields

runStateT :: s -> m (a, s)

module MoreMonoids

module StrategyPrelude

module StrategyInfix

module OverloadingTheme

module TraversalTheme

module FlowTheme

module FixpointTheme

module KeyholeTheme

module NameTheme

module PathTheme

module EffectTheme

module ContainerTheme

module RefactoringTheme

module MetricsTheme