StrategyPrelude

Description

This module is part of StrategyLib, a library of functional strategy combinators, including combinators for generic traversal. This module is basically a wrapper for the strategy primitives plus some extra basic strategy combinators that can be defined immediately in terms of the primitive ones.

Documentation

module StrategyPrimitives

Useful defaults for strategy update (see StrategyPrimitives.adhocTU and StrategyPrimitives.adhocTP).

idTP :: GHC.Base.Monad m => StrategyPrimitives.TP m

Type-preserving identity. Returns the incoming term without change.

failTP :: Control.Monad.MonadPlus m => StrategyPrimitives.TP m

Type-preserving failure. Always fails, independent of the incoming term. Uses Control.Monad.MonadPlus to model partiality.

failTU :: Control.Monad.MonadPlus m => StrategyPrimitives.TU a m

Type-unifying failure. Always fails, independent of the incoming term. Uses Control.Monad.MonadPlus to model partiality.

constTU :: GHC.Base.Monad m => a -> StrategyPrimitives.TU a m

Type-unifying constant strategy. Always returns the argument value a, independent of the incoming term.

compTU :: GHC.Base.Monad m => m a -> StrategyPrimitives.TU a m

Type-unifying monadic constant strategy. Always performs the argument computation a, independent of the incoming term. This is a monadic variation of StrategyPrelude.constTU.

Lift a function to a strategy type with failure as default

monoTP :: (TermRep.Term a, Control.Monad.MonadPlus m) => (a -> m a) -> StrategyPrimitives.TP m

Apply the monomorphic, type-preserving argument function, if its input type matches the input term's type. Otherwise, fail.

monoTU :: (TermRep.Term a, Control.Monad.MonadPlus m) => (a -> m b) -> StrategyPrimitives.TU b m

Apply the monomorphic, type-unifying argument function, if its input type matches the input term's type. Otherwise, fail.

Function composition

dotTU :: GHC.Base.Monad m => (a -> b) -> StrategyPrimitives.TU a m -> StrategyPrimitives.TU b m

Sequential ccomposition of monomorphic function and type-unifying strategy. In other words, after the type-unifying strategy s has been applied, the monomorphic function f is applied to the resulting value.

op2TU :: GHC.Base.Monad m => (a -> b -> c) -> StrategyPrimitives.TU a m -> StrategyPrimitives.TU b m -> StrategyPrimitives.TU c m

Parallel combination of two type-unifying strategies with a binary combinator. In other words, the values resulting from applying the type-unifying strategies are combined to a final value by applying the combinator o.

Reduce a strategy's performance to its effects

voidTP :: GHC.Base.Monad m => StrategyPrimitives.TP m -> StrategyPrimitives.TU GHC.Tuple.() m

Reduce a type-preserving strategy to a type-unifying one that ignores its result term and returns void, but retains its monadic effects.

voidTU :: GHC.Base.Monad m => StrategyPrimitives.TU u m -> StrategyPrimitives.TU GHC.Tuple.() m

Reduce a type-unifying strategy to a type-unifying one that ignores its result value and returns void, but retains its monadic effects.

Shape test combinators

con :: Control.Monad.MonadPlus m => StrategyPrimitives.TP m

Test for constant term, i.e. having no subterms.

com :: Control.Monad.MonadPlus m => StrategyPrimitives.TP m

Test for compound term, i.e. having at least one subterm.