TraversalTheme

Description

This module is part of StrategyLib, a library of functional strategy combinators, including combinators for generic traversal. This module defines traversal schemes. Such schemes have formed the core of StrategyLib since its first release. The portfolio as it stands now captures part of the design in the paper ... Polymorphic Symphony.

Recursive traversal

Full traversals

full_tdTP :: GHC.Base.Monad m => StrategyPrimitives.TP m -> StrategyPrimitives.TP m

Full type-preserving traversal in top-down order.

full_buTP :: GHC.Base.Monad m => StrategyPrimitives.TP m -> StrategyPrimitives.TP m

Full type-preserving traversal in bottom-up order.

full_tdTU :: (GHC.Base.Monad m, Data.Monoid.Monoid a) => StrategyPrimitives.TU a m -> StrategyPrimitives.TU a m

Full type-unifying traversal in top-down order.

Traversals with stop conditions

stop_tdTP :: Control.Monad.MonadPlus m => StrategyPrimitives.TP m -> StrategyPrimitives.TP m

Top-down type-preserving traversal that is cut of below nodes where the argument strategy succeeds.

stop_tdTU :: (Control.Monad.MonadPlus m, Data.Monoid.Monoid a) => StrategyPrimitives.TU a m -> StrategyPrimitives.TU a m

Top-down type-unifying traversal that is cut of below nodes where the argument strategy succeeds.

Single hit traversal

once_tdTP :: Control.Monad.MonadPlus m => StrategyPrimitives.TP m -> StrategyPrimitives.TP m

Top-down type-preserving traversal that performs its argument strategy at most once.

once_tdTU :: Control.Monad.MonadPlus m => StrategyPrimitives.TU a m -> StrategyPrimitives.TU a m

Top-down type-unifying traversal that performs its argument strategy at most once.

once_buTP :: Control.Monad.MonadPlus m => StrategyPrimitives.TP m -> StrategyPrimitives.TP m

Bottom-up type-preserving traversal that performs its argument strategy at most once.

once_buTU :: Control.Monad.MonadPlus m => StrategyPrimitives.TU a m -> StrategyPrimitives.TU a m

Bottom-up type-unifying traversal that performs its argument strategy at most once.

Traversal with environment propagation

once_peTU

Arguments

:: Control.Monad.MonadPlus m

=> e
initial environment

-> (e -> StrategyPrimitives.TU e m)
environment modification at downward step

-> (e -> StrategyPrimitives.TU a m)
extraction of value, dependent on environment

-> StrategyPrimitives.TU a m

Top-down type-unifying traversal with propagation of an environment.

One-layer traversal

Defined versions of some primitive one-layer traversal combinators

anyTP' :: Control.Monad.MonadPlus m => StrategyPrimitives.TP m -> StrategyPrimitives.TP m

Use StrategyPrimitives.anyTP instead.

someTP' :: Control.Monad.MonadPlus m => StrategyPrimitives.TP m -> StrategyPrimitives.TP m

Use StrategyPrimitives.someTP instead.

Recursive completion of one-layer traversal

all_recTU

Arguments

:: (Data.Monoid.Monoid a, GHC.Base.Monad m)

=> (t -> StrategyPrimitives.TU a m -> StrategyPrimitives.TU a m)
binary strategy combinator

-> t
argument strategy

-> StrategyPrimitives.TU a m
result strategy

Recursive completion of full type-preserving one-layer traverasal

one_recTU

Arguments

:: Control.Monad.MonadPlus m

=> (t -> StrategyPrimitives.TU a m -> StrategyPrimitives.TU a m)
binary strategy combinator

-> t
argument strategy

-> StrategyPrimitives.TU a m
result strategy

Recursive completion of type-preserving one-layer traversal that succeeds exactly once.

Overloading and synonyms

Overloaded schemes for traversal

full_td :: OverloadingTheme.StrategyMonoid s m => s m -> s m

Full top-down traversal (overloaded between StrategyPrimitives.TU and StrategyPrimitives.TP).

once_td :: OverloadingTheme.StrategyPlus s m => s m -> s m

One-hit top-down traversal (overloaded between StrategyPrimitives.TU and StrategyPrimitives.TP).

once_bu :: OverloadingTheme.StrategyPlus s m => s m -> s m

One-hit bottom-up traversal (overloaded between StrategyPrimitives.TU and StrategyPrimitives.TP).

once_pe :: OverloadingTheme.StrategyPlus s m => (e -> s m) -> (e -> StrategyPrimitives.TU e m) -> e -> s m

One-hit top-down traversal with environment propagation (overloaded between StrategyPrimitives.TU and StrategyPrimitives.TP).

Some synonyms for convenience

topdown :: GHC.Base.Monad m => StrategyPrimitives.TP m -> StrategyPrimitives.TP m

See TraversalTheme.full_tdTP.

crush :: (GHC.Base.Monad m, Data.Monoid.Monoid u) => StrategyPrimitives.TU u m -> StrategyPrimitives.TU u m

See TraversalTheme.full_tdTU.

collect :: GHC.Base.Monad m => StrategyPrimitives.TU [a] m -> StrategyPrimitives.TU [a] m

Type-specialised version of TraversalTheme.crush, which works with lists instead of any arbitrary monoid.

select :: Control.Monad.MonadPlus m => StrategyPrimitives.TU u m -> StrategyPrimitives.TU u m

See TraversalTheme.once_tdTU.

selectenv :: Control.Monad.MonadPlus m => e -> (e -> StrategyPrimitives.TU e m) -> (e -> StrategyPrimitives.TU a m) -> StrategyPrimitives.TU a m

See TraversalTheme.once_peTU.